TTSServer/requirements.txt
# TTS Server Dependencies

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# Configuration
pyyaml>=6.0

# Audio processing
numpy>=1.21.0

# TTS Backends
edge-tts>=6.1.0

# Optional: Qwen TTS (uncomment to use)
# qwen-tts>=0.1.0
# torch>=2.0.0
